types corrected
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@6507 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtbase/IRevisionController.h
@@ -33,7 +33,7 @@ public:
 		{
 		}
 
-		int revision;
+		qlonglong revision;
 		QDateTime timestamp;
 		QString user;
 		QString comment;
